+++ service-a/src/main/java/io/x99/service_a/Main.java
@@ -1,4 +1,4 @@
-package io.x99;
+package io.x99.service_a;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
