+++ service-a/src/main/java/io/x99/service_a/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package io.x99.repository;
+package io.x99.service_a.repository;
 
-import io.x99.model.ProductEntity;
+import io.x99.service_a.model.ProductEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
